add manifest file to package
