+++ src/packages/core/property-editor/manifests.ts
@@ -1,4 +1,3 @@
 import { manifests as propertyEditorSchemaManifests } from './schemas/manifests.js';
-import { manifests as propertyEditorUIManifests } from './uis/manifests.js';
 
-export const manifests = [...propertyEditorSchemaManifests, ...propertyEditorUIManifests];
+export const manifests = [...propertyEditorSchemaManifests];
